LPS-64255 move AutoBatchPSUtil out of portal-impl upgrade and into portal-kernel dao jdbc where it makes more sense, and where it can be more easily reused
